// frontend/src/shared/types/appModules.ts
export type AppModule =
  | 'agenda'
  | 'staff'
  | 'inventory'
  | 'pos'
  | 'commissions'
  | 'services'
  | 'tables'
  | 'digital_menu'
  | 'subscriptions'
  | 'access_control'
  | 'progress_tracking'
  | 'assets_management'
  | 'contracts'
  | 'projects'
  | 'tasks'
  | 'kitchen_display'
  | 'accounting'
  | 'ecommerce_storefront'
  | 'shopping_cart'
  | 'order_management'
  | 'shipping_tracking'
  | 'multi_vendor';